parse_args implemented

+++ philosopher/srcs/time/timing.c
@@ -6,7 +6,7 @@
 /*   By: kclaudan <kclaudan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/13 12:00:00 by philosopher       #+#    #+#             */
-/*   Updated: 2025/03/13 15:59:29 by kclaudan         ###   ########.fr       */
+/*   Updated: 2025/03/13 17:22:47 by kclaudan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@
  *
  * Utilise gettimeofday pour obtenir le temps actuel en millisecondes.
  */
-long	get_timestamp_ms(void)
+int	get_timestamp_ms(void)
 {
 	/* À implémenter */
 	return (0);
@@ -33,7 +33,7 @@ long	get_timestamp_ms(void)
  *
  * Calcule la différence entre le timestamp actuel et le timestamp de départ.
  */
-long	get_elapsed_time(t_simulation *sim)
+int	get_elapsed_time(t_simulation *sim)
 {
 	/* À implémenter */
 	return (0);
@@ -47,7 +47,7 @@ long	get_elapsed_time(t_simulation *sim)
  * Utilise usleep avec de petits intervalles pour une meilleure précision.
  * Évite la dérive temporelle sur de longues périodes.
  */
-void	precise_sleep(long duration_ms)
+void	precise_sleep(int duration_ms)
 {
 	/* À implémenter */
 }
